backend/requirements.txt
flask
django
python-dotenv
groq
cohere